Major Refactor: Now have example/radial
Rendering of the Blocks is now the preserve of the library user.
This allow the two examples to style the blocks very differently.

Added the concept of a "bounding rectangle", limiting the placement of
new blocks into a smaller region.

example/radial: A new example to show off the coralling of large block
into the central region.